order amount validated

+++ ondoc/api/v1/auth/views.py
@@ -1465,7 +1465,7 @@ class TransactionViewSet(viewsets.GenericViewSet):
 
         order_obj = Order.objects.select_for_update().filter(pk=response.get("orderId")).first()
         convert_cod_to_prepaid = False
-        if int(order_obj.amount) != int(response.get('txAmount')):
+        if float(order_obj.amount) != float(response.get('txAmount')):
             raise Exception('amount mismatched for the transaction')
         try:
             # if order_obj and response and order_obj.is_cod_order and order_obj.get_deal_price_without_coupon <= Decimal(response.get('txAmount')):
@@ -1658,7 +1658,7 @@ class TransactionViewSet(viewsets.GenericViewSet):
 
                 order_obj = Order.objects.select_for_update().filter(pk=order_id).first()
                 convert_cod_to_prepaid = False
-                if int(order_obj.amount) != int(response.get('txAmount')):
+                if float(order_obj.amount) != float(response.get('txAmount')):
                     raise Exception('amount mismatched for the transaction')
                 try:
                     # if order_obj and response and order_obj.is_cod_order and order_obj.get_deal_price_without_coupon <= Decimal(response.get('txAmount')):
